backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
gunicorn==21.2.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.9.2
email-validator==2.1.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
requests==2.31.0
anthropic==0.39.0
apscheduler==3.10.4
python-multipart==0.0.12
psycopg2-binary==2.9.10
python-slugify==8.0.4
beautifulsoup4==4.12.3
html5lib==1.1
asyncpg==0.29.0
greenlet==3.0.3
aiosqlite==0.20.0